Problem 4 - Brute Force
A pretty bad solution for the fourth problem. It iterates through the first 100 numbers (meaning it checks 10000 numbers in the worst case scenario) and then splits each digit into a vector, copies the vector and reverses it and then checks if the vectors are equal. It's a very pythonic way of approaching this problem, so I will attack it from another, more C++-ish, angle and try again.