Problem 6 - Brute Force
This method uses a for loop, giving it linear time complexity.

Whilst thinking about this problem, I noticed that the formula deduced by Gauss to work out the sum of a series of numbers (n^2 + n) / 2 follows the triangle numbers; 1, 3, 6, 10, 15, etc. I researched the equivalent of this in 3 dimensions and came across the formula for square pyramidal numbers, whihc would make this solution constant time complexity