Problem 7
I'm relatively happy with this solution, after playing about with trying to get a sieve of Eratosthenes working! It uses an array of booleans and a std::vector to store the primes. The only thing that I am not happy with is that I have to define a max value (which I called k_max) and that number is completely arbitrary, meaning there is a lot of unused space.

I also discovered a new keyword! memset, which can prefill a memory location. Will be useful in future!