Problem 13 - 100 50 digit numbers
My solution is pretty fast, which I am happy with, but it is not very space efficient as I am creating std::vector<std::vector<T>> all over the place. It treats the 50 numbers as arrays of digits (which are characters for simplicity) and then adds them and carries as if a human was doing column addition in arithmetic.